feat: Update MNIST model to use uint8 inputs
This commit modifies the MNIST model to use uint8 inputs instead of int64 inputs. The `gen.py` file in the `mnist_uint8` directory has been updated to change the data type of the input tensor from int64 to uint8. This change ensures that the model receives the correct input data type for performing inference on MNIST images.