refactor: Update MNIST model to use smaller convolutional filters
This commit updates the MNIST model in the `gen.py` file to use smaller convolutional filters. The `conv1` filter size is changed from 3x3 to 5x5, and the `conv2` filter size is changed from 32x32 to 8x8. This change aims to improve the model's performance by reducing the number of parameters and increasing the receptive field of the filters.